refac: kind icon
