+++ pages/Entities.tsx
@@ -46,4 +46,3 @@ const styles = StyleSheet.create({
         bottom: 0,
     }
 });
-
